+++ NEWS.md
@@ -1,7 +1,28 @@
+<!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
+
 # mlr3spatial 0.1.2.9000
 
-- fix: The `$data()` method of `DataBackendRaster` sometimes failed when not adjacent cells were read.
-- refactor: compatibility to `terra` update.
+## Bug fixes
+
+* write result to correct cell
+
+* test: predictions are written to the right cells
+
+* renamed geometry column in DataBackendVector (#49)
+
+* The `$data()` method of `DataBackendRaster` should not fail anymore when non-adjacent cells are read
+
+## Features
+
+* add `as_task_classif` and as_task`_regr (#53)
+
+* add training task to backends without response
+
+* refactor: compatibility to `{terra}` update.
+
+## Documentation
+
+* update benchmark vignette
 
 # mlr3spatial 0.1.2